+++ packages/ckeditor5-dev-env/tests/release-tools/tasks/releasesubrepositories.js
@@ -34,20 +34,24 @@ const filesToRecoverAfterTest = [
 ];
 
 describe( 'dev-env/release-tools/tasks', function() {
-	this.timeout( 10 * 1000 ); // 10 sec * 1000 ms
+	this.timeout( 15 * 1000 ); // 15 sec * 1000 ms
 
-	let releaseSubRepositories, sandbox, stubs;
+	let releaseSubRepositories, sandbox, stubs, executedCommand;
 
 	beforeEach( () => {
+		executedCommand = [];
+
 		for ( const filePath of filesToRecoverAfterTest ) {
 			originalFilesContent[ filePath ] = fs.readFileSync( filePath, 'utf-8' );
 		}
 
 		for ( const packageName of Object.keys( packagesPaths ) ) {
 			process.chdir( packagesPaths[ packageName ] );
+
 			// In order to be sure that we are testing the clean repository.
 			exec( `rm -rf ${ path.join( packagesPaths[ packageName ], '.git' ) }` );
 			exec( 'git init --quiet' );
+			exec( `git remote add origin https://github.com/ckeditor/ckeditor5-test-${ packageName }.git` );
 
 			if ( process.env.CI ) {
 				exec( 'git config user.email "ckeditor5@ckeditor.com"' );
@@ -78,16 +82,14 @@ describe( 'dev-env/release-tools/tasks', function() {
 				error: sandbox.spy()
 			},
 			displaySkippedPackages: sandbox.stub(),
-			releaseRepository: sandbox.stub(),
-			generateChangelogForSinglePackage: sandbox.stub(),
-			validatePackageToRelease: sandbox.stub()
+			validatePackageToRelease: sandbox.stub(),
+			createGithubRelease: sandbox.stub()
 		};
 
-		mockery.registerMock( '../utils/releaserepository', stubs.releaseRepository );
-		mockery.registerMock( './generatechangelogforsinglepackage', stubs.generateChangelogForSinglePackage );
 		mockery.registerMock( '../utils/displayskippedpackages', stubs.displaySkippedPackages );
 		mockery.registerMock( '../utils/cli', stubs.cli );
 		mockery.registerMock( '../utils/validatepackagetorelease', stubs.validatePackageToRelease );
+		mockery.registerMock( '../utils/creategithubrelease', stubs.createGithubRelease );
 
 		sandbox.stub( path, 'join' ).callsFake( ( ...chunks ) => chunks.join( '/' ) );
 
@@ -95,7 +97,10 @@ describe( 'dev-env/release-tools/tasks', function() {
 			'@ckeditor/ckeditor5-dev-utils': {
 				tools: {
 					shExec( command, options ) {
-						if ( command === 'git pull && git push' ) {
+						executedCommand.push( command );
+
+						// Does nothing with the remote commands.
+						if ( command === 'git pull' || command === 'git push' || command.startsWith( 'npm publish' ) ) {
 							return '';
 						}
 
@@ -170,7 +175,7 @@ describe( 'dev-env/release-tools/tasks', function() {
 				.then( () => {
 					expect( stubs.logger.info.calledOnce ).to.equal( true );
 					expect( stubs.logger.info.firstCall.args[ 0 ] ).to.equal(
-						'Aborted due to user\'s no confirmation.'
+						'Releasing has been aborted.'
 					);
 				} );
 		} );
@@ -231,8 +236,6 @@ describe( 'dev-env/release-tools/tasks', function() {
 				skipGithub: true
 			} ) );
 
-			stubs.releaseRepository.returns( Promise.resolve() );
-
 			return releaseSubRepositories( options )
 				.then( () => {
 					expect( stubs.validatePackageToRelease.callCount ).to.equal( 1 );
@@ -243,85 +246,7 @@ describe( 'dev-env/release-tools/tasks', function() {
 				} );
 		} );
 
-		it( 'generates changelog if dependencies of package have been changed', () => {
-			const options = {
-				cwd: mainPackagePath,
-				packages: 'packages'
-			};
-
-			stubs.cli.confirmRelease.returns( Promise.resolve( true ) );
-
-			makeCommit( 'alpha', 'Feature: This is an initial commit.' );
-			makeChangelog( 'alpha', [
-				'Changelog',
-				'=========',
-				'',
-				'## 0.1.0 (2017-06-08)',
-				'',
-				'### Features',
-				'',
-				'* This is an initial commit.',
-				''
-			].join( '\n' ) );
-			makeCommit( 'alpha', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
-
-			stubs.generateChangelogForSinglePackage.onFirstCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'beta', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.2.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'beta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
-
-						resolve();
-					} );
-				} ) );
-
-			stubs.generateChangelogForSinglePackage.onFirstCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'delta', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.4.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'delta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
-
-						resolve();
-					} );
-				} ) );
-
-			stubs.validatePackageToRelease.returns( [] );
-			stubs.cli.configureReleaseOptions.returns( Promise.resolve( {
-				skipNpm: true,
-				skipGithub: true
-			} ) );
-
-			stubs.releaseRepository.returns( Promise.resolve() );
-
-			return releaseSubRepositories( options )
-				.then( () => {
-					expect( stubs.generateChangelogForSinglePackage.calledTwice ).to.equal( true );
-					expect( stubs.generateChangelogForSinglePackage.firstCall.args[ 0 ] ).to.equal( '0.2.1' );
-					expect( stubs.generateChangelogForSinglePackage.secondCall.args[ 0 ] ).to.equal( '0.4.1' );
-				} );
-		} );
-
-		it( 'releases packages', () => {
+		it( 'releases packages (npm and github)', () => {
 			const options = {
 				cwd: mainPackagePath,
 				packages: 'packages'
@@ -357,105 +282,57 @@ describe( 'dev-env/release-tools/tasks', function() {
 
 			stubs.cli.confirmRelease.returns( Promise.resolve( true ) );
 
-			stubs.generateChangelogForSinglePackage.onFirstCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'delta', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.4.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'delta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+			const token = 'fooToken';
 
-						resolve();
-					} );
-				} ) );
-
-			stubs.generateChangelogForSinglePackage.onSecondCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'epsilon', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.5.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'epsilon', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
-
-						resolve();
-					} );
-				} ) );
-
-			stubs.cli.configureReleaseOptions.returns( Promise.resolve( {
-				skipNpm: true,
-				skipGithub: true
-			} ) );
+			stubs.cli.configureReleaseOptions.resolves( {
+				token,
+				skipNpm: false,
+				skipGithub: false
+			} );
 
 			stubs.validatePackageToRelease.returns( [] );
-			stubs.releaseRepository.returns( Promise.resolve() );
+			stubs.createGithubRelease.resolves();
 
 			return releaseSubRepositories( options )
 				.then( () => {
-					expect( stubs.releaseRepository.callCount ).to.equal( 4 );
+					expect( executedCommand.length ).to.equal( 14 );
+					expect( stubs.createGithubRelease.callCount ).to.equal( 2 );
 
 					// Alpha
-					expect( stubs.releaseRepository.getCall( 0 ).args[ 0 ] ).to.deep.equal( {
-						skipGithub: true,
-						skipNpm: true,
-						token: undefined,
-						version: '0.1.0',
-						changes: [
-							'### Features',
-							'',
-							'* This is an initial commit.'
-						].join( '\n' )
+					expect( executedCommand[ 0 ], 'Alpha diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 1 ], 'Alpha add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 2 ], 'Alpha commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 6 ], 'Alpha version' ).to.equal( 'npm version 0.1.0 --message "Release: v0.1.0."' );
+					expect( executedCommand[ 8 ], 'Alpha publish' ).to.equal( 'npm publish --access=public' );
+					expect( executedCommand[ 10 ], 'Alpha push' ).to.equal( 'git push' );
+					expect( executedCommand[ 12 ], 'Alpha remote' ).to.equal( 'git remote get-url origin --push' );
+					expect( stubs.createGithubRelease.getCall( 0 ).args[ 0 ] ).to.equal( token );
+					expect( stubs.createGithubRelease.getCall( 0 ).args[ 1 ] ).to.deep.equal( {
+						repositoryOwner: 'ckeditor',
+						repositoryName: 'ckeditor5-test-alpha',
+						version: 'v0.1.0',
+						description: '### Features\n\n* This is an initial commit.'
 					} );
 
 					// Beta
-					expect( stubs.releaseRepository.getCall( 1 ).args[ 0 ] ).to.deep.equal( {
-						skipGithub: true,
-						skipNpm: true,
-						token: undefined,
-						version: '0.2.1',
-						changes: [
-							'### Fix',
-							'',
-							'* Some fix.'
-						].join( '\n' )
-					} );
-
-					// Delta
-					expect( stubs.releaseRepository.getCall( 2 ).args[ 0 ] ).to.deep.equal( {
-						skipGithub: true,
-						skipNpm: true,
-						token: undefined,
-						version: '0.4.1',
-						changes: 'Internal changes only (updated dependencies, documentation, etc.).'
-					} );
-
-					// Epsilon
-					expect( stubs.releaseRepository.getCall( 3 ).args[ 0 ] ).to.deep.equal( {
-						skipGithub: true,
-						skipNpm: true,
-						token: undefined,
-						version: '0.5.1',
-						changes: 'Internal changes only (updated dependencies, documentation, etc.).'
+					expect( executedCommand[ 3 ], 'Beta diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 4 ], 'Beta add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 5 ], 'Beta commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 7 ], 'Beta version' ).to.equal( 'npm version 0.2.1 --message "Release: v0.2.1."' );
+					expect( executedCommand[ 9 ], 'Beta publish' ).to.equal( 'npm publish --access=public' );
+					expect( executedCommand[ 11 ], 'Beta push' ).to.equal( 'git push' );
+					expect( executedCommand[ 13 ], 'Beta remote' ).to.equal( 'git remote get-url origin --push' );
+					expect( stubs.createGithubRelease.getCall( 1 ).args[ 0 ] ).to.equal( token );
+					expect( stubs.createGithubRelease.getCall( 1 ).args[ 1 ] ).to.deep.equal( {
+						repositoryOwner: 'ckeditor',
+						repositoryName: 'ckeditor5-test-beta',
+						version: 'v0.2.1',
+						description: '### Fix\n\n* Some fix.'
 					} );
 				} );
 		} );
 
-		it( 'must not release any package if any error occured', () => {
+		it( 'releases packages (github only)', () => {
 			const options = {
 				cwd: mainPackagePath,
 				packages: 'packages'
@@ -491,49 +368,160 @@ describe( 'dev-env/release-tools/tasks', function() {
 
 			stubs.cli.confirmRelease.returns( Promise.resolve( true ) );
 
-			stubs.generateChangelogForSinglePackage.onFirstCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'delta', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.4.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'delta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+			const token = 'fooToken';
 
-						resolve();
+			stubs.cli.configureReleaseOptions.resolves( {
+				token,
+				skipNpm: true,
+				skipGithub: false
+			} );
+
+			stubs.validatePackageToRelease.returns( [] );
+			stubs.createGithubRelease.resolves();
+
+			return releaseSubRepositories( options )
+				.then( () => {
+					expect( executedCommand.length ).to.equal( 12 );
+					expect( stubs.createGithubRelease.callCount ).to.equal( 2 );
+
+					// Alpha
+					expect( executedCommand[ 0 ], 'Alpha diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 1 ], 'Alpha add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 2 ], 'Alpha commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 6 ], 'Alpha version' ).to.equal( 'npm version 0.1.0 --message "Release: v0.1.0."' );
+					expect( executedCommand[ 8 ], 'Alpha push' ).to.equal( 'git push' );
+					expect( executedCommand[ 10 ], 'Alpha remote' ).to.equal( 'git remote get-url origin --push' );
+					expect( stubs.createGithubRelease.getCall( 0 ).args[ 0 ] ).to.equal( token );
+					expect( stubs.createGithubRelease.getCall( 0 ).args[ 1 ] ).to.deep.equal( {
+						repositoryOwner: 'ckeditor',
+						repositoryName: 'ckeditor5-test-alpha',
+						version: 'v0.1.0',
+						description: '### Features\n\n* This is an initial commit.'
 					} );
-				} ) );
 
-			stubs.generateChangelogForSinglePackage.onSecondCall()
-				.returns( new Promise( resolve => {
-					// Without this timeout, the promise is resolving before starting the test and `getPackagesToRelease()`
-					// returns an invalid values.
-					setTimeout( () => {
-						makeChangelog( 'epsilon', [
-							'Changelog',
-							'=========',
-							'',
-							'## [0.5.1](https://githubn.com/ckeditor) (2017-06-08)',
-							'',
-							'Internal changes only (updated dependencies, documentation, etc.).',
-							''
-						].join( '\n' ) );
-						makeCommit( 'epsilon', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
-
-						resolve();
+					// Beta
+					expect( executedCommand[ 3 ], 'Beta diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 4 ], 'Beta add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 5 ], 'Beta commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 7 ], 'Beta version' ).to.equal( 'npm version 0.2.1 --message "Release: v0.2.1."' );
+					expect( executedCommand[ 9 ], 'Beta push' ).to.equal( 'git push' );
+					expect( executedCommand[ 11 ], 'Beta remote' ).to.equal( 'git remote get-url origin --push' );
+					expect( stubs.createGithubRelease.getCall( 1 ).args[ 0 ] ).to.equal( token );
+					expect( stubs.createGithubRelease.getCall( 1 ).args[ 1 ] ).to.deep.equal( {
+						repositoryOwner: 'ckeditor',
+						repositoryName: 'ckeditor5-test-beta',
+						version: 'v0.2.1',
+						description: '### Fix\n\n* Some fix.'
 					} );
-				} ) );
+				} );
+		} );
+
+		it( 'releases packages (npm only)', () => {
+			const options = {
+				cwd: mainPackagePath,
+				packages: 'packages'
+			};
+
+			makeCommit( 'alpha', 'Feature: This is an initial commit.' );
+			makeChangelog( 'alpha', [
+				'Changelog',
+				'=========',
+				'',
+				'## 0.1.0 (2017-06-08)',
+				'',
+				'### Features',
+				'',
+				'* This is an initial commit.',
+				''
+			].join( '\n' ) );
+			makeCommit( 'alpha', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+
+			makeCommit( 'beta', 'Fix: Some fix.' );
+			makeChangelog( 'beta', [
+				'Changelog',
+				'=========',
+				'',
+				'## [0.2.1](https://github.com/ckeditor) (2017-06-08)',
+				'',
+				'### Fix',
+				'',
+				'* Some fix.',
+				''
+			].join( '\n' ) );
+			makeCommit( 'beta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+
+			stubs.cli.confirmRelease.returns( Promise.resolve( true ) );
+
+			stubs.cli.configureReleaseOptions.resolves( {
+				skipNpm: false,
+				skipGithub: true
+			} );
+
+			stubs.validatePackageToRelease.returns( [] );
+			stubs.createGithubRelease.resolves();
+
+			return releaseSubRepositories( options )
+				.then( () => {
+					expect( executedCommand.length ).to.equal( 12 );
+					expect( stubs.createGithubRelease.called ).to.equal( false );
+
+					// Alpha
+					expect( executedCommand[ 0 ], 'Alpha diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 1 ], 'Alpha add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 2 ], 'Alpha commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 6 ], 'Alpha version' ).to.equal( 'npm version 0.1.0 --message "Release: v0.1.0."' );
+					expect( executedCommand[ 8 ], 'Alpha publish' ).to.equal( 'npm publish --access=public' );
+					expect( executedCommand[ 10 ], 'Alpha push' ).to.equal( 'git push' );
+
+					// Beta
+					expect( executedCommand[ 3 ], 'Beta diff' ).to.equal( 'git diff --name-only package.json' );
+					expect( executedCommand[ 4 ], 'Beta add to commit' ).to.equal( 'git add package.json' );
+					expect( executedCommand[ 5 ], 'Beta commit' ).to.equal( 'git commit -m "Internal: Updated dependencies. [skip ci]"' );
+					expect( executedCommand[ 7 ], 'Beta version' ).to.equal( 'npm version 0.2.1 --message "Release: v0.2.1."' );
+					expect( executedCommand[ 9 ], 'Beta publish' ).to.equal( 'npm publish --access=public' );
+					expect( executedCommand[ 11 ], 'Beta push' ).to.equal( 'git push' );
+				} );
+		} );
+
+		it( 'must not release any package if any error occurred', () => {
+			const options = {
+				cwd: mainPackagePath,
+				packages: 'packages'
+			};
+
+			makeCommit( 'alpha', 'Feature: This is an initial commit.' );
+			makeChangelog( 'alpha', [
+				'Changelog',
+				'=========',
+				'',
+				'## 0.1.0 (2017-06-08)',
+				'',
+				'### Features',
+				'',
+				'* This is an initial commit.',
+				''
+			].join( '\n' ) );
+			makeCommit( 'alpha', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+
+			makeCommit( 'beta', 'Fix: Some fix.' );
+			makeChangelog( 'beta', [
+				'Changelog',
+				'=========',
+				'',
+				'## [0.2.1](https://github.com/ckeditor) (2017-06-08)',
+				'',
+				'### Fix',
+				'',
+				'* Some fix.',
+				''
+			].join( '\n' ) );
+			makeCommit( 'beta', 'Docs: Changelog. [skip ci]', [ 'CHANGELOG.md' ] );
+
+			stubs.cli.confirmRelease.returns( Promise.resolve( true ) );
 
 			stubs.validatePackageToRelease.returns( [] );
 			stubs.validatePackageToRelease.onFirstCall().returns( [
-				'Not on master or master is not clean.'
+				'Not on master.'
 			] );
 			stubs.validatePackageToRelease.onSecondCall().returns( [
 				'Cannot find changelog entry for version 0.2.1.'
@@ -541,13 +529,12 @@ describe( 'dev-env/release-tools/tasks', function() {
 
 			return releaseSubRepositories( options )
 				.then( () => {
-					expect( stubs.validatePackageToRelease.callCount ).to.equal( 4 );
-					expect( stubs.releaseRepository.called ).to.equal( false );
+					expect( stubs.validatePackageToRelease.callCount ).to.equal( 2 );
 
 					expect( stubs.logger.error.callCount ).to.equal( 5 );
 					expect( stubs.logger.error.getCall( 0 ).args[ 0 ] ).to.equal( 'Releasing has been aborted due to errors.' );
 					expect( stubs.logger.error.getCall( 1 ).args[ 0 ] ).to.equal( '## @ckeditor/alpha' );
-					expect( stubs.logger.error.getCall( 2 ).args[ 0 ] ).to.equal( '* Not on master or master is not clean.' );
+					expect( stubs.logger.error.getCall( 2 ).args[ 0 ] ).to.equal( '* Not on master.' );
 					expect( stubs.logger.error.getCall( 3 ).args[ 0 ] ).to.equal( '## @ckeditor/beta' );
 					expect( stubs.logger.error.getCall( 4 ).args[ 0 ] ).to.equal( '* Cannot find changelog entry for version 0.2.1.' );
 				} );