+++ packages/ckeditor5-dev-env/tests/release-tools/utils/getsubpackagespaths.js
@@ -91,5 +91,106 @@ describe( 'dev-env/release-tools/utils', () => {
 			expect( pathsCollection.skipped.size ).to.equal( 1 );
 			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-utils' ) ).to.equal( true );
 		} );
+
+		it( 'allows ignoring specified packages (specified as string)', () => {
+			getDirectoriesStub.returns( [
+				'ckeditor5-core',
+				'ckeditor5-engine',
+				'ckeditor5-utils'
+			] );
+
+			const options = {
+				cwd: '/tmp',
+				packages: 'packages',
+				skipPackages: '@ckeditor/ckeditor5-u*'
+			};
+
+			getPackageJsonStub.onFirstCall().returns( { name: '@ckeditor/ckeditor5-core' } );
+			getPackageJsonStub.onSecondCall().returns( { name: '@ckeditor/ckeditor5-engine' } );
+			getPackageJsonStub.onThirdCall().returns( { name: '@ckeditor/ckeditor5-utils' } );
+
+			const pathsCollection = getSubPackagesPaths( options );
+
+			expect( pathsCollection.packages ).to.be.instanceof( Set );
+			expect( pathsCollection.packages.size ).to.equal( 2 );
+
+			expect( pathsCollection.skipped ).to.be.instanceof( Set );
+			expect( pathsCollection.skipped.size ).to.equal( 1 );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-utils' ) ).to.equal( true );
+		} );
+
+		it( 'allows restricting the scope for packages', () => {
+			getDirectoriesStub.returns( [
+				'ckeditor5-core',
+				'ckeditor5-engine',
+				'ckeditor5-utils',
+				'ckeditor5-build-classic',
+				'ckeditor5-build-inline',
+			] );
+
+			const options = {
+				cwd: '/tmp',
+				packages: 'packages',
+				scope: '@ckeditor/ckeditor5-build-*',
+				skipPackages: []
+			};
+
+			getPackageJsonStub.onCall( 0 ).returns( { name: '@ckeditor/ckeditor5-core' } );
+			getPackageJsonStub.onCall( 1 ).returns( { name: '@ckeditor/ckeditor5-engine' } );
+			getPackageJsonStub.onCall( 2 ).returns( { name: '@ckeditor/ckeditor5-utils' } );
+			getPackageJsonStub.onCall( 3 ).returns( { name: '@ckeditor/ckeditor5-build-classic' } );
+			getPackageJsonStub.onCall( 4 ).returns( { name: '@ckeditor/ckeditor5-build-inline' } );
+
+			const pathsCollection = getSubPackagesPaths( options );
+
+			expect( pathsCollection.packages ).to.be.instanceof( Set );
+			expect( pathsCollection.packages.size ).to.equal( 2 );
+			expect( pathsCollection.packages.has( '/tmp/packages/ckeditor5-build-classic' ) ).to.equal( true );
+			expect( pathsCollection.packages.has( '/tmp/packages/ckeditor5-build-inline' ) ).to.equal( true );
+
+			expect( pathsCollection.skipped ).to.be.instanceof( Set );
+			expect( pathsCollection.skipped.size ).to.equal( 3 );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-core' ) ).to.equal( true );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-engine' ) ).to.equal( true );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-utils' ) ).to.equal( true );
+		} );
+
+		it( 'allows restricting the scope for packages and works fine with "skipPackages" option', () => {
+			getDirectoriesStub.returns( [
+				'ckeditor5-core',
+				'ckeditor5-engine',
+				'ckeditor5-utils',
+				'ckeditor5-build-classic',
+				'ckeditor5-build-inline',
+			] );
+
+			const options = {
+				cwd: '/tmp',
+				packages: 'packages',
+				scope: '@ckeditor/ckeditor5-build-*',
+				skipPackages: [
+					'@ckeditor/ckeditor5-build-inline'
+				]
+			};
+
+			getPackageJsonStub.onCall( 0 ).returns( { name: '@ckeditor/ckeditor5-core' } );
+			getPackageJsonStub.onCall( 1 ).returns( { name: '@ckeditor/ckeditor5-engine' } );
+			getPackageJsonStub.onCall( 2 ).returns( { name: '@ckeditor/ckeditor5-utils' } );
+			getPackageJsonStub.onCall( 3 ).returns( { name: '@ckeditor/ckeditor5-build-classic' } );
+			getPackageJsonStub.onCall( 4 ).returns( { name: '@ckeditor/ckeditor5-build-inline' } );
+
+			const pathsCollection = getSubPackagesPaths( options );
+
+			expect( pathsCollection.packages ).to.be.instanceof( Set );
+			expect( pathsCollection.packages.size ).to.equal( 1 );
+			expect( pathsCollection.packages.has( '/tmp/packages/ckeditor5-build-classic' ) ).to.equal( true );
+
+			expect( pathsCollection.skipped ).to.be.instanceof( Set );
+			expect( pathsCollection.skipped.size ).to.equal( 4 );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-core' ) ).to.equal( true );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-engine' ) ).to.equal( true );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-utils' ) ).to.equal( true );
+			expect( pathsCollection.skipped.has( '/tmp/packages/ckeditor5-build-inline' ) ).to.equal( true );
+		} );
 	} );
 } );